+++ tests/emailpassword/test_emailverify.py
@@ -936,7 +936,7 @@ async def test_the_generate_token_api_with_valid_input_and_then_remove_token(dri
     start_st()
 
     version = await Querier.get_instance().get_api_version()
-    assert version in ('2.9', '2.10', '2.11', '2.12')
+    assert version in ('2.9', '2.10', '2.11', '2.12', '2.13')
 
     response_1 = sign_up_request(
         driver_config_client,
@@ -974,7 +974,7 @@ async def test_the_generate_token_api_with_valid_input_verify_and_then_unverify_
     start_st()
 
     version = await Querier.get_instance().get_api_version()
-    assert version in ('2.9', '2.10', '2.11', '2.12')
+    assert version in ('2.9', '2.10', '2.11', '2.12', '2.13')
 
     response_1 = sign_up_request(
         driver_config_client,