+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add delete email (`delete_email_for_user`) and phone number (`delete_phone_number_for_user`) functions for passwordless and thirdpartypasswordless recipe
 - Upgrade and freeze pyright version
 - Rename `compare_version` to `get_max_version` for readability
+- Add user metadata recipe
 
 ## [0.6.4] - 2022-04-11
 - bug fix in `default_create_and_send_custom_email` for emailverification recipe where we were not sending the email if env var was not set.